A single cut: SDL2 and SDL3 headers conflict and SDL_Scancode is a
transitive dependency of core, debug, both frontends and 7 suites, so
there was no incremental path.

Found by doing the work, not predicted by the design note: SDL3 delivers
mouse wheel and motion as float and the two need opposite handling —
wheel must read integer_y (reading the float truncates every sub-detent
scroll of a high-resolution wheel to zero), motion has no integer twin
so the remainder is carried between events.

GamepadHost had zero coverage: every joystick row drove JoystickDispatcher
one layer below with synthetic events, so all of them would have stayed
green on a host that opened the wrong device or none. Now driven by a real
SDL3 virtual device.

The audio-underrun row's missing-capture branch is now fail_row, not
skip_row — SDL3 renaming SDL_DISKAUDIOFILE would otherwise have retired
the project's only underrun test silently.

Ubuntu 24.04 has no libsdl3, so package-deb vendors a pinned SDL 3.4.16
built static, sha256 cross-checked against two hosts before unpacking.
The static link drops audio backends libsdl2-2.0-0 used to pull in, so
the .deb now Depends on libpipewire | libpulse | libasound as
alternatives.

unit 8409 -> 8421, 115 suites, 0 fail, 0 skip; FUSE 1356/1356;
regression 144/144; package-test 20 pass, 1 pre-existing skip.

# Conflicts:
#	doc/testing/TRACEABILITY-MATRIX.md
#	test/SUBSYSTEM-TESTS-STATUS.md

+++ src/platform/sdl_app.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 #include <memory>
 #include <string>
 #include "sdl_display.h"
@@ -103,8 +103,8 @@ private:
     bool mouse_captured_ = false;
     void set_mouse_captured(bool on);
 
-    // Joystick / gamepad host adapter — wires SDL_CONTROLLER* events into
-    // the emulator's Joystick and owns the SDL_GameController lifecycle for
+    // Joystick / gamepad host adapter — wires SDL_EVENT_GAMEPAD_* events into
+    // the emulator's Joystick and owns the SDL_Gamepad lifecycle for
     // the two Next pad headers. Created in init() once emulator_ is bound.
     // Closes G42 (JOY-WIRE-02/03/04); Task 79 moved the controller
     // open/close/route logic into the shared GamepadHost.
@@ -153,7 +153,8 @@ private:
     // --speed (see set_speed_percent()); 1.0 = real time.
     double      speed_multiplier_ = 1.0;
     // SDL_GetTicks() of the last present, for the >1x present throttle.
-    uint32_t    last_present_ms_ = 0;
+    // Uint64-wide because SDL3's SDL_GetTicks() is (SDL2's was Uint32).
+    uint64_t    last_present_ms_ = 0;
 
     // Pending --rzx-play / --rzx-record, started at the top of run().
     std::string rzx_play_file_;
